- make math::bit_* functions template - cleanup - fix 0xFA instruction incorrectly pass the memory address to reg a

+++ src/cpu/Register16.cpp
@@ -34,7 +34,7 @@ gameboy::cpu::Register16& gameboy::cpu::Register16::operator--()
 
 gameboy::cpu::Register16& gameboy::cpu::Register16::operator+=(uint16_t value)
 {
-    *this = *this + value;
+    *this = static_cast<uint16_t>(*this + value);
     return *this;
 }
 
@@ -52,7 +52,7 @@ gameboy::cpu::Register16& gameboy::cpu::Register16::operator+=(const memory::Add
 
 gameboy::cpu::Register16& gameboy::cpu::Register16::operator-=(uint16_t value)
 {
-    *this = *this - value;
+    *this = static_cast<uint16_t>(*this - value);
     return *this;
 }
